Move twitter models to twitter/models
